json/yaml merger
